+++ internal/bloblang/field/resolver.go
@@ -51,6 +51,7 @@ func (q QueryResolver) ResolveString(index int, msg Message, escaped, legacy boo
 		Index:    index,
 		MsgBatch: msg,
 		Legacy:   legacy,
+		NewMeta:  msg.Get(index).Metadata(),
 	}.WithValueFunc(func() *interface{} {
 		if jObj, err := msg.Get(index).JSON(); err == nil {
 			return &jObj
@@ -65,6 +66,7 @@ func (q QueryResolver) ResolveBytes(index int, msg Message, escaped, legacy bool
 		Index:    index,
 		MsgBatch: msg,
 		Legacy:   legacy,
+		NewMeta:  msg.Get(index).Metadata(),
 	}.WithValueFunc(func() *interface{} {
 		if jObj, err := msg.Get(index).JSON(); err == nil {
 			return &jObj